Replacing Headless UI with custom implementation.

+++ Api/Migrations/SimpleProductOrderServiceDbContextModelSnapshot.cs
@@ -69,7 +69,7 @@ namespace Api.Migrations
                     b.Property<int>("Status")
                         .HasColumnType("int");
 
-                    b.Property<float>("Total")
+                    b.Property<decimal>("Total")
                         .HasColumnType("real");
 
                     b.HasKey("Id");
@@ -140,7 +140,7 @@ namespace Api.Migrations
                         .HasMaxLength(255)
                         .HasColumnType("nvarchar(255)");
 
-                    b.Property<float>("Price")
+                    b.Property<decimal>("Price")
                         .HasColumnType("real");
 
                     b.HasKey("Id");
@@ -170,7 +170,7 @@ namespace Api.Migrations
                         .HasMaxLength(255)
                         .HasColumnType("nvarchar(255)");
 
-                    b.Property<float>("Price")
+                    b.Property<decimal>("Price")
                         .HasColumnType("real");
 
                     b.Property<int?>("ProductId")
