+++ Application/Api/Orders/List/DTOs/ListOrdersRequestDTO.cs
@@ -2,7 +2,7 @@ namespace Application.Api.Orders.List.DTOs;
 
 public class ListOrdersRequestDTO
 {
-    public ListOrdersRequestDTO(float? minTotal, float? maxTotal, string? status, DateTime? createdBefore, DateTime? createdAfter)
+    public ListOrdersRequestDTO(decimal? minTotal, decimal? maxTotal, string? status, DateTime? createdBefore, DateTime? createdAfter)
     {
         MinTotal = minTotal;
         MaxTotal = maxTotal;
@@ -11,8 +11,8 @@ public class ListOrdersRequestDTO
         CreatedAfter = createdAfter;
     }
 
-    public float? MinTotal { get; set; }
-    public float? MaxTotal { get; set; }
+    public decimal? MinTotal { get; set; }
+    public decimal? MaxTotal { get; set; }
     public string? Status { get; set; }
     public DateTime? CreatedBefore { get; set; }
     public DateTime? CreatedAfter { get; set; }
