+++ Application/Api/Products/List/DTOs/ListProductsRequestDTO.cs
@@ -4,8 +4,8 @@ public class ListProductsRequestDTO
 {
     public ListProductsRequestDTO(
         string? name, 
-        float? minPrice, 
-        float? maxPrice, 
+        decimal? minPrice, 
+        decimal? maxPrice, 
         string? description, 
         DateTime? createdBefore, 
         DateTime? createdAfter)
@@ -19,8 +19,8 @@ public class ListProductsRequestDTO
     }
 
     public string? Name { get; set; }
-    public float? MinPrice { get; set; }
-    public float? MaxPrice { get; set; }
+    public decimal? MinPrice { get; set; }
+    public decimal? MaxPrice { get; set; }
     public string? Description { get; set; }
     public DateTime? CreatedBefore { get; set; }
     public DateTime? CreatedAfter { get; set; }
