+++ Application/Api/Products/List/Handlers/ListProductsQuery.cs
@@ -6,7 +6,7 @@ namespace Application.Api.Products.List.Handlers;
 
 public class ListProductsQuery : IRequest<OneOf<ListProductsResult, List<PlainApplicationError>>>
 {
-    public ListProductsQuery(string? name, float? minPrice, float? maxPrice, string? description, DateTime? createdBefore, DateTime? createdAfter)
+    public ListProductsQuery(string? name, decimal? minPrice, decimal? maxPrice, string? description, DateTime? createdBefore, DateTime? createdAfter)
     {
         Name = name;
         MinPrice = minPrice;
@@ -17,8 +17,8 @@ public class ListProductsQuery : IRequest<OneOf<ListProductsResult, List<PlainAp
     }
 
     public string? Name { get; set; }
-    public float? MinPrice { get; set; }
-    public float? MaxPrice { get; set; }
+    public decimal? MinPrice { get; set; }
+    public decimal? MaxPrice { get; set; }
     public string? Description { get; set; }
     public DateTime? CreatedBefore { get; set; }
     public DateTime? CreatedAfter { get; set; }
